+++ packages/blocto-sdk/src/providers/blocto.ts
@@ -13,7 +13,8 @@ class BloctoProvider implements EIP1193Provider {
   isConnecting = false;
   appId?: string;
 
-  eventListeners: { [key: string]: Array<(arg?: any) => void> } = {};
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  eventListeners: { [key: string]: Array<(arg: any) => void> } = {};
 
   sessionKey = KEY_SESSION;
 
@@ -27,16 +28,19 @@ class BloctoProvider implements EIP1193Provider {
     });
   }
 
-  protected formatAndSetSessionAccount(address: Record<string, string> = {}) {
+  protected formatAndSetSessionAccount(
+    address: Record<string, string> = {}
+  ): void {
     this.session.accounts = Object.keys(address).reduce<
       Record<string, string[]>
     >((initial, current) => {
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       initial[current] = [address![current]];
       return initial;
     }, {});
   }
 
-  protected tryRetrieveSessionFromStorage(chain: string) {
+  protected tryRetrieveSessionFromStorage(chain: string): void {
     // load previous connected state
     const session = getItemWithExpiry<Session>(this.sessionKey, {});
     const sessionCode = session && session.code;